FinancialManager/requirements.txt
plaid-python==16.2.0
slack-sdk==3.27.2
APScheduler==3.10.4
python-dotenv==1.0.1
PyYAML==6.0.1
